test: address Avg fixture review feedback
