+++ src/DynamicData.Tests/AggregationTests/AvgFixture.cs
@@ -1,0 +1,3 @@
+namespace DynamicData.Tests.AggregationTests;
+
+public partial class AvgFixture;
